+++ util/fsutilprogress/progress.go
@@ -1,8 +1,10 @@
 package fsutilprogress
 
 import (
+	"fmt"
 	"path"
 	"sync"
+	"time"
 
 	"github.com/dustin/go-humanize"
 	"github.com/earthly/earthly/conslogging"
@@ -16,13 +18,18 @@ type ProgressCallback interface {
 }
 
 type progressCallback struct {
-	console     conslogging.ConsoleLogger
-	mutex       sync.Mutex
-	pathPrefix  string
-	numStats    int
-	numSent     int
-	numReceived int
-	filesize    map[string]int
+	console           conslogging.ConsoleLogger
+	mutex             sync.Mutex
+	pathPrefix        string
+	numStats          int
+	numSent           int
+	numReceived       int
+	bytesSent         int
+	bytesReceived     int
+	filesize          map[string]int
+	lastUpdate        time.Time
+	lastBytesSent     int
+	lastBytesReceived int
 }
 
 // New returns a new verbose progress callback for use with fsutil
@@ -53,9 +60,10 @@ func (s *progressCallback) Verbose(relPath string, status fsutil.VerboseProgress
 	case fsutil.StatusSent:
 		s.console.VerbosePrintf("sent data for %s (%s)\n", fullPath, humanize.Bytes(uint64(numBytes)))
 		s.numSent++
+		s.bytesSent += numBytes
 	case fsutil.StatusReceiving:
 		s.filesize[fullPath] += numBytes
-		//ignore
+		s.bytesReceived += numBytes
 	case fsutil.StatusReceived:
 		if numBytes == 0 {
 			numBytes = s.filesize[fullPath]
@@ -69,4 +77,36 @@ func (s *progressCallback) Verbose(relPath string, status fsutil.VerboseProgress
 	default:
 		s.console.Warnf("unhandled progress status %v (path=%s, numBytes=%d)\n", status, fullPath, numBytes)
 	}
+
+	// display a summary every 15 seconds
+	now := time.Now()
+	d := now.Sub(s.lastUpdate)
+	if d > time.Second*15 {
+		if s.numSent > 0 {
+			var transferRate string
+			if !s.lastUpdate.IsZero() {
+				transferRate = fmt.Sprintf("; transfer rate: %s/s", humanize.Bytes(uint64(float64(s.bytesSent-s.lastBytesSent)/d.Seconds())))
+			}
+			s.console.Printf("sent %s (%s)%s\n", humanize.Bytes(uint64(s.bytesSent)), puralize(s.numSent, "file"), transferRate)
+		} else {
+			s.console.Printf("sent %s)\n", puralize(s.numStats, "file stat"))
+		}
+		if s.numReceived > 0 {
+			var transferRate string
+			if !s.lastUpdate.IsZero() {
+				transferRate = fmt.Sprintf("; transfer rate: %s/s", humanize.Bytes(uint64(float64(s.bytesReceived-s.lastBytesReceived)/d.Seconds())))
+			}
+			s.console.Printf("received %s (%s)%s\n", humanize.Bytes(uint64(s.bytesReceived)), puralize(s.numReceived, "file"), transferRate)
+		}
+		s.lastUpdate = now
+		s.lastBytesSent = s.bytesSent
+		s.lastBytesReceived = s.bytesReceived
+	}
+}
+
+func puralize(n int, suffix string) string {
+	if n == 1 {
+		return "1 " + suffix
+	}
+	return fmt.Sprintf("%d %ss", n, suffix)
 }